+++ app/src/main/java/pt/android/cvnotes/domain/repository/NoteRepository.kt
@@ -5,7 +5,7 @@ import pt.android.cvnotes.domain.model.Note
 
 interface NoteRepository {
     fun getNotes(): Flow<List<Note>>
-    suspend fun getNoteById(id: Int): Note?
+    suspend fun getNoteById(id: Long): Note?
     suspend fun insertNote(note: Note)
     suspend fun deleteNote(note: Note)
 }
