+++ app/src/main/java/pt/android/cvnotes/domain/use_case/note/GetNoteById.kt
@@ -6,7 +6,7 @@ import pt.android.cvnotes.domain.repository.NoteRepository
 class GetNoteById(
     private val noteRepository: NoteRepository
 ) {
-    suspend operator fun invoke(id: Int): Note? {
+    suspend operator fun invoke(id: Long): Note? {
         return noteRepository.getNoteById(id)
     }
 }
